requirements.txt
tqdm==4.50.2
imutils==0.5.3
uvicorn==0.13.3
numpy==1.19.1
pytesseract==0.3.7
opencv_python==4.5.1.48
absl_py==0.11.0
fastapi==0.63.0
Pillow==8.2.0
PyYAML==5.4.1
rich==10.1.0
